--- security-analysis/contingency-screening-security-analysis/src/main/resources/com/powsybl/contingencyscreening/security/analysis/reports.properties
@@ -1,5 +1,0 @@
-contingencyScreeningSecurityAnalysis = Running contingency screening security analysis on network '${networkId}'.
-contingencyScreeningSecurityAnalysisTotalContingencies = Total number of contingencies to process: ${count}.
-contingencyScreeningSecurityAnalysisFirstPassStarted = First pass started using provider '${providerName}'.
-contingencyScreeningSecurityAnalysisSecondPassRequired = Second pass required for ${count} contingency(ies).
-contingencyScreeningSecurityAnalysisSecondPassStarted = Second pass started using provider '${providerName}'.
