--- security-analysis/contingency-screening-security-analysis/src/main/resources/com/powsybl/contingencyscreening/security/analysis/reports_fr.properties
@@ -1,5 +1,0 @@
-contingencyScreeningSecurityAnalysis = Exécution de l'analyse de sécurité avec filtrage des contingences sur le réseau '${networkId}'.
-contingencyScreeningSecurityAnalysisTotalContingencies = Nombre total de contingences à traiter : ${count}.
-contingencyScreeningSecurityAnalysisFirstPassStarted = Premier passage démarré avec le fournisseur '${providerName}'.
-contingencyScreeningSecurityAnalysisSecondPassRequired = Deuxième passage requis pour ${count} contingence(s).
-contingencyScreeningSecurityAnalysisSecondPassStarted = Deuxième passage démarré avec le fournisseur '${providerName}'.
