--- security-analysis/contingency-screening-security-analysis/src/test/resources/config.yml
@@ -1,3 +1,0 @@
-contingency-screening-security-analysis-parameters:
-  firstProviderName: LoadFlow
-  secondProviderName: DynaFlow
