+++ security-analysis/pom.xml
@@ -25,7 +25,7 @@
 
     <modules>
         <module>security-analysis-api</module>
-        <module>contingency-screening-security-analysis</module>
+        <module>twopass-security-analysis</module>
     </modules>
 
 </project>
