+++ security-analysis/twopass-security-analysis/src/main/resources/com/powsybl/twopasssecurity/security/analysis/reports.properties
@@ -1,0 +1,5 @@
+twoPassSecurityAnalysis = Running two-pass security analysis on network '${networkId}'.
+twoPassSecurityAnalysisTotalContingencies = Total number of contingencies to process: ${count}.
+twoPassSecurityAnalysisFirstPassStarted = First pass started using provider '${providerName}'.
+twoPassSecurityAnalysisSecondPassRequired = Second pass required for ${count} contingency(ies).
+twoPassSecurityAnalysisSecondPassStarted = Second pass started using provider '${providerName}'.
