+++ security-analysis/twopass-security-analysis/src/main/resources/com/powsybl/twopasssecurity/security/analysis/reports_fr.properties
@@ -1,0 +1,5 @@
+twoPassSecurityAnalysis = Exécution de l'analyse de sécurité avec filtrage des contingences sur le réseau '${networkId}'.
+twoPassSecurityAnalysisTotalContingencies = Nombre total de contingences à traiter : ${count}.
+twoPassSecurityAnalysisFirstPassStarted = Premier passage démarré avec le fournisseur '${providerName}'.
+twoPassSecurityAnalysisSecondPassRequired = Deuxième passage requis pour ${count} contingence(s).
+twoPassSecurityAnalysisSecondPassStarted = Deuxième passage démarré avec le fournisseur '${providerName}'.
