+++ security-analysis/twopass-security-analysis/src/test/java/com/powsybl/twopasssecurity/security/analysis/parameters/TwoPassSecurityAnalysisParametersTest.java
@@ -1,0 +1,96 @@
+/**
+ * Copyright (c) 2026, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+package com.powsybl.twopasssecurity.security.analysis.parameters;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+/**
+ * @author Riad Benradi {@literal <riad.benradi_externe at rte-france.com>}
+ */
+class TwoPassSecurityAnalysisParametersTest {
+    private TwoPassSecurityAnalysisParameters twoPassSecurityAnalysisParameters;
+
+    @BeforeEach
+    void setUp() {
+        twoPassSecurityAnalysisParameters = new TwoPassSecurityAnalysisParameters();
+    }
+
+    @Test
+    void testDefaultValues() {
+        assertNull(twoPassSecurityAnalysisParameters.getFirstProviderName());
+        assertNull(twoPassSecurityAnalysisParameters.getSecondProviderName());
+    }
+
+    @Test
+    void testSetBothProviders() {
+        twoPassSecurityAnalysisParameters.setFirstProviderName("OpenLoadFlow")
+                .setSecondProviderName("DynaFlow");
+
+        assertEquals("OpenLoadFlow", twoPassSecurityAnalysisParameters.getFirstProviderName());
+        assertEquals("DynaFlow", twoPassSecurityAnalysisParameters.getSecondProviderName());
+    }
+
+    @Test
+    void testGetName() {
+        assertEquals("twopass-security-analysis-parameters", twoPassSecurityAnalysisParameters.getName());
+    }
+
+    @Test
+    void testSetGetFirstProviderName() {
+        twoPassSecurityAnalysisParameters.setFirstProviderName("OpenLoadFlow");
+        assertEquals("OpenLoadFlow", twoPassSecurityAnalysisParameters.getFirstProviderName());
+    }
+
+    @Test
+    void testSetGetSecondProviderName() {
+        twoPassSecurityAnalysisParameters.setSecondProviderName("Dynaflow");
+        assertEquals("Dynaflow", twoPassSecurityAnalysisParameters.getSecondProviderName());
+    }
+
+    @Test
+    void testCompleteConfiguration() {
+        twoPassSecurityAnalysisParameters.setFirstProviderName("OpenLoadFlow")
+                .setSecondProviderName("Dynaflow");
+        assertEquals("OpenLoadFlow", twoPassSecurityAnalysisParameters.getFirstProviderName());
+        assertEquals("Dynaflow", twoPassSecurityAnalysisParameters.getSecondProviderName());
+    }
+
+    @Test
+    void testUpdate() {
+        TwoPassSecurityAnalysisParameters parameters = new TwoPassSecurityAnalysisParameters();
+
+        // Initial state
+        assertNull(parameters.getFirstProviderName());
+        assertNull(parameters.getSecondProviderName());
+
+        // Update both
+        Map<String, String> properties = new HashMap<>();
+        properties.put("firstProviderName", "static-flow");
+        properties.put("secondProviderName", "dynamic-sim");
+
+        parameters.update(properties);
+
+        assertEquals("static-flow", parameters.getFirstProviderName());
+        assertEquals("dynamic-sim", parameters.getSecondProviderName());
+
+        // Partial update: only second provider
+        Map<String, String> partialProperties = new HashMap<>();
+        partialProperties.put("secondProviderName", "new-dynamic-sim");
+
+        parameters.update(partialProperties);
+
+        // First one should remain unchanged
+        assertEquals("static-flow", parameters.getFirstProviderName());
+        assertEquals("new-dynamic-sim", parameters.getSecondProviderName());
+    }
+}
