+++ security-analysis/twopass-security-analysis/src/test/resources/config.yml
@@ -1,0 +1,3 @@
+twopass-security-analysis-parameters:
+  firstProviderName: LoadFlow
+  secondProviderName: DynaFlow
